refactor: 챌린지 사이클 API 리팩토링
