+++ frontend/src/apis/cycleApi/type.ts
@@ -1,0 +1,9 @@
+import { Cycle } from 'commonType';
+
+export type PostCycleProps = Pick<Cycle, 'challengeId' | 'startTime'>;
+
+export type PostCycleProgressProps = Pick<Cycle, 'cycleId'>;
+
+export interface PostCycleProgressResponse {
+  progressCount: number;
+}
