+++ frontend/src/apis/index.ts
@@ -1,3 +1,4 @@
 export * from 'apis/authApi';
-export * from 'apis/challengeApi';
+export * from 'apis/cycleApi';
 export * from 'apis/oAuthApi';
+export * from 'apis/challengeApi';
