+++ frontend/src/components/ChallengeList/index.tsx
@@ -1,5 +1,5 @@
 import { ChallengeInfo } from './type';
-import { useGetAllChallenges } from 'apis/challengeApi';
+import { useGetAllChallenges } from 'apis/cycleApi';
 import styled, { css } from 'styled-components';
 
 import { FlexBox, ChallengeItem } from 'components';
