+++ frontend/src/mocks/handlers/challenge.js
@@ -19,17 +19,17 @@ export const challenge = [
   rest.get(`${DEV_BASE_URL}/cycles/me?status=inProgress`, (req, res, ctx) => {
     return res(ctx.status(200), ctx.json(cycleData));
   }),
-  // 3. 챌린지 사이클의 진척도 증가(POST)
+  // 4. 챌린지 사이클의 진척도 증가(POST)
   rest.post(`${DEV_BASE_URL}/cycles/:cycleId/progress`, (req, res, ctx) => {
     const { cycleId } = req.params;
     cycleData[cycleId - 1].progressCount++;
     return res(ctx.status(200), ctx.json({ progressCount: 2 }));
   }),
-  // 4. 모든 챌린지 조회(GET)
+  // 5. 모든 챌린지 조회(GET)
   rest.get(`${DEV_BASE_URL}/challenges`, (req, res, ctx) => {
     return res(ctx.status(200), ctx.json(challengeData));
   }),
-  // 5. 아이디로 사이클 조회(GET)
+  // 7. 아이디로 사이클 조회(GET)
   rest.post(`${DEV_BASE_URL}${API_PATH.CYCLES_ID}`, (req, res, ctx) => {
     return res(ctx.status(200), ctx.json(cycleData[0]));
   }),
